+++ api/src/Routes/index.js
@@ -1,10 +1,11 @@
-const {Router} = require('express');
+const { Router } = require("express");
 const router = Router();
-const {checkEmail} = require('../Controllers/get.js')
+const ProductRoutes = require("./Products/products");
+const UserRoutes = require("./user/users");
+const MoviesRoutes = require("./Movies/movies");
 
-router.use('/SignUp', checkEmail)
-
-
-
+router.use("/users", UserRoutes);
+router.use("/products", ProductRoutes);
+router.use("/movies", MoviesRoutes);
 
 module.exports = router;
